Fix router
