+++ src/App.js
@@ -1,13 +1,13 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { BrowserRouter as Routes, Router, Route } from 'react-router-dom';
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-
+import { fetchPostsRequestDragons } from './redux/Dragons/dragons';
 import './App.css';
 import Rocket from './component/rockets/rocket';
 import Mission from './component/missions/mission';
 import MyProfile from './component/my-profile/profile';
-import { fetchPostsRequestDragons } from './redux/Dragons/dragons';
-import Dragons from './component/Dragons/Dragons';
+
+/* import Dragons from './component/Dragons/Dragons'; */
 import Header from './component/Header/Header';
 
 function App() {
@@ -19,15 +19,14 @@ function App() {
   return (
     <Router>
       <div className="App">
+
         <Header />
         <Routes>
-          <Route path="/rockets" element={<Rocket />} />
-          <Route path="/mission" element={<Mission />} />
-          <Route path="/Dragons" element={<Dragons />} />
-          <Route path="/profile" element={<MyProfile />} />
-
-          <Route path="/" exact element={<Rocket />} />
+          <Route path="/Mission" element={<Mission />} />
+          <Route path="/MyProfile" element={<MyProfile />} />
+          <Route path="/" element={<Rocket />} />
         </Routes>
+
       </div>
     </Router>
   );
