+++ src/redux/API/rocketAPI.js
@@ -13,16 +13,3 @@ export default async function FetchRocketsData() {
     throw new Error('Can not retrieve data from the response');
   }
 }
-
-const dragonsUrl = 'https://api.spacexdata.com/v3/dragons';
-
-const getData = async (url) => {
-  try {
-    const response = await fetch(url);
-    return response.json();
-  } catch (error) {
-    throw new Error(error.message);
-  }
-};
-
-export const fetchDragons = () => getData(dragonsUrl);
