+++ src/redux/configureStore.js
@@ -2,7 +2,6 @@ import { combineReducers, createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
 import missionReducer, { fetchMissionSuccess } from './missions/mission';
-
 import profileReducer from './profiles/profile';
 import rocketReducer from './rockets/rockets';
 import dragonReducer from './Dragons/dragons';
